polymarket-telegram-bot/requirements.txt
python-telegram-bot==13.15
requests==2.32.3
PyYAML==6.0.2
python-dotenv==1.0.1
web3==6.0.0
pytest==7.4.3
